mqtt-pms5003/requirements.txt
paho-mqtt>=1.3.1
pms5003>=0.0.5 
